requirements.txt
asgiref==3.8.1
Cython==3.0.10
Django==5.0.7
django-mssql-backend==2.8.1
joblib==1.4.2
mssql-django==1.5
numpy==1.25.2
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pmdarima==2.0.4
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.1
statsmodels==0.14.2
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
